[ARM] Sabrelite board support improved.
The Sabrelite can now be selected in the menuconfig, and the specific board
initialization source skeleton is done.
It also adds the clock management for the i.MX6Q, which is from the Linux
kernel and can be largely improved and split as drivers ; and simplified GPC,
timers, PM, and all clock management functions.

The i.MX6Q GPT driver is specific to the i.MX6Q. The device tree GPT definition
compatibility is set to with "fsl,imx6q-gpt", instead of "freescale,gpt-timer".

The SOC name and revision will be displayed when asking for the board
information through the Xvisor command line.