[ARM] i.MX6 CCM device added in the device tree, required for LP clocking.
As the CCM was not declared in the device tree, the associated driver, in
arch/arm/cpu/board/common/imx/clk-imx6q.c was not initialized. Thus, the Low
Power Manager clocking did not set the low power wait mode as clocked, state in
which a WFI set the processor to. In consequence, when the first proc_do_idle
was called, the JTAG probe lost its control over the processor.
The scheduling was functioning though, as the EPIT timer was running.