[DRIVERS] Adding MXC GPIO driver.
As the Generic GPIO driver is already ported on Xvisor, using it will help the
port of the i.MX GPIO driver, especially the mapping between the IRQ and the
GPIO pin indexing.

Xvisor device tree format is slightly different from Linux's one.
The IRQ should be written as plain number, and not without specifying the
sharing flag nor the interrupt type.
This although can be written with these specifiers (modifying the driver
accordingly), but it would break Xvisor device tree format.

Also, the IRQ number is the IRQ line, without the 32 base IRQ offset.