[DRIVERS] SabreLite SPI NOR flash driver port.
This drivers requires a bus id to be set in the parent device node, otherwise,
the bus id will be set as the bus id 0.
The M25P80 can now be used as both a character device, and a block device in
Xvisor.
The M25P80 NOR flash and similar character driver has been implemented,
following the new character device API, which allows to use the "vmm_chardev_*"
operations for this flash.