[DRIVERS] Correcting and improving the i.MX GPIO base offset.
The GPIO base offset was set to 0. It must be left to -1, for the GPIO lib to
find an empty GPIO description place, or correctly set.
Instead of using the GPIO lib offset finding, the GPIO base as been set from
the controller physical address, defined in the device tree.
GPIO group 1 is set in the "gpio_desc" structure from 0 to 31, GPIO group 2
from 32 to 63, GPIO group 3 from 64 to 95 and so on.