[CORE] Extended IRQ support added.
This add extended IRQ support in Xvisor.
The aim is to have a light and simple Linux IRQ domain like, without modifying
the existing Xvisor IRQs.

More precisely, it maps multiple IRQ on an hard IRQ.
Those extended IRQ number are defined above the upper bound of
CONFIG_HOST_IRQ_COUNT.

This helps, for example, GPIO drivers to have multiple IRQ on a single hardware
one.
In order to be used, the driver must first map a number of extended IRQs with
"vmm_host_extirq_map", retrieving a "vmm_host_extirqs" structure pointer, which
must be saved.
When the GPIO driver is asked for the IRQ number associated to a specific pin,
it just have to call "vmm_host_extirq_get_irq" with its "vmm_host_extirqs"
saved structure to retrive the extended IRQ, which can be used as a regular
IRQ.

For now, the total number of extended IRQs and extended IRQ groups are set in
the configuration, and cannot be unmapped, but should be latter dynamic (using
for example a bitmap allocation).